chore: Update layout styles for main content and add AdSense ads
Adjust the layout styles for the main content in the `layout.module.css` file. Increase the left margin to accommodate a left ad and add a right margin for a right ad. Also, add new CSS classes `.adContainer` and `.rightAdContainer` to position the ads using fixed positioning.

In the `layout.tsx` file, import the `AdSenseAd` component and add two instances of it within the layout. One instance is placed in the `.adContainer` div on the left side, and the other instance is placed in the `.rightAdContainer` div on the right side.

In the `home.tsx` file, import the `AdSenseAd` component and add an instance of it before the heading to display a horizontal ad.

Added jenkins to docker workflow

added apiUrl to components

added API_URL to Dockerfile